Airflow/requirements.txt
packaging<21.0
elasticsearch-dbapi
aiohttp
alembic
async-lru
attrs
Babel
beautifulsoup4
cachelib
certifi
cffi
charset-normalizer
click
colorlog
cryptography
dash
dash-core-components
dash-html-components
dash-table
decorator
dnspython
Flask
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-SQLAlchemy
Flask-WTF
gunicorn
idna
importlib-metadata
itsdangerous
Jinja2
jsonschema
Mako
Markdown
MarkupSafe
marshmallow
numpy
networkx
pandas
pendulum
plotly
psutil
pycparser
Pygments
python-dateutil
python-slugify
pytz
requests
SQLAlchemy
sqlparse
tenacity
Werkzeug
WTForms
python-dotenv
